// src/types/evenement.tsx
export interface EvenementDocument {
    id?: string;
    title?: string;
    date?: string;
    description?: string;
    facebook?: string;
    instagram?: string;
    tiktok?: string;
    youtube?: string;
    twitter?: string;
    image?: string | null;
    userId?: string;
    creation_date?: unknown;
}